refactor and add unit test for AuditRequestParams()
