+++ internal/federationdomain/endpoints/auth/auth_handler_test.go
@@ -30,12 +30,12 @@ import (
 
 	supervisorfake "go.pinniped.dev/generated/latest/client/supervisor/clientset/versioned/fake"
 	"go.pinniped.dev/generated/latest/client/supervisor/clientset/versioned/typed/config/v1alpha1"
+	"go.pinniped.dev/internal/auditid"
 	"go.pinniped.dev/internal/authenticators"
 	"go.pinniped.dev/internal/federationdomain/csrftoken"
 	"go.pinniped.dev/internal/federationdomain/endpoints/jwks"
 	"go.pinniped.dev/internal/federationdomain/oidc"
 	"go.pinniped.dev/internal/federationdomain/oidcclientvalidator"
-	"go.pinniped.dev/internal/federationdomain/requestlogger"
 	"go.pinniped.dev/internal/federationdomain/stateparam"
 	"go.pinniped.dev/internal/federationdomain/storage"
 	"go.pinniped.dev/internal/here"
@@ -4118,7 +4118,7 @@ func TestAuthorizationEndpoint(t *testing.T) { //nolint:gocyclo
 		if test.customPasswordHeader != nil {
 			req.Header.Set("Pinniped-Password", *test.customPasswordHeader)
 		}
-		req, _ = requestlogger.NewRequestWithAuditID(req, func() string { return "fake-audit-id" })
+		req, _ = auditid.NewRequestWithAuditID(req, func() string { return "fake-audit-id" })
 		rsp := httptest.NewRecorder()
 
 		subject.ServeHTTP(rsp, req)
