+++ internal/federationdomain/requestlogger/request_logger.go
@@ -11,10 +11,6 @@ import (
 	"slices"
 	"time"
 
-	"github.com/google/uuid"
-	"k8s.io/apimachinery/pkg/types"
-	apisaudit "k8s.io/apiserver/pkg/apis/audit"
-	"k8s.io/apiserver/pkg/audit"
 	"k8s.io/apiserver/pkg/endpoints/responsewriter"
 	"k8s.io/utils/clock"
 
@@ -23,31 +19,6 @@ import (
 	"go.pinniped.dev/internal/httputil/requestutil"
 	"go.pinniped.dev/internal/plog"
 )
-
-// NewRequestWithAuditID is public for use in unit tests. Production code should use WithAuditID().
-func NewRequestWithAuditID(r *http.Request, newAuditIDFunc func() string) (*http.Request, string) {
-	ctx := audit.WithAuditContext(r.Context())
-	r = r.WithContext(ctx)
-
-	auditID := newAuditIDFunc()
-	audit.WithAuditID(ctx, types.UID(auditID))
-
-	return r, auditID
-}
-
-func WithAuditID(handler http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// Add a randomly generated request ID to the context for this request.
-		r, auditID := NewRequestWithAuditID(r, func() string {
-			return uuid.New().String()
-		})
-
-		// Send the Audit-ID response header.
-		w.Header().Set(apisaudit.HeaderAuditID, auditID)
-
-		handler.ServeHTTP(w, r)
-	})
-}
 
 func WithHTTPRequestAuditLogging(handler http.Handler, auditLogger plog.AuditLogger, auditInternalPathsCfg supervisor.AuditInternalPaths) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
